xray_flask/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
numpy>=1.21.0
